chore/integration tests: print RNG seed
Print RNG seed for each test. It doesn't look very pretty when running
all tests in parallel as you can see the same seed printed multiple
times and the order is meaningless.

I still think it is worth printing as it may make the difference between
us being able to answer a question or not (as the seed would let us
reproduce the specific scenario).

Picking functionality over esthaetics :)