Interrupt the test if consensus complete or broken
Even in a correct execution the nodes only come to consensus after an
unknown number of steps. This means we don't know how many steps we need
to generate in order for the test to pass. In order to give the nodes
enough time, we will now generate a lot of steps in advance, but
interrupt the execution as soon as consensus is reached; or, if there is
a bug, if we detect that nodes have disagreeing orders of transactions.