fix/parsec: resolve add_event issues
This ensures that when a new node is handling its initial request from a
peer which will contain the entire graph, it places all these events
before its own ones in the 'events_order' list.  So now, when forming
its response, it skips all these events which it has just been given and
only includes its own.

It also fixes an associated issue whereby we were sending events to a
peer which included each of the last ancestors of that peer's latest
event, rather than all the ones newer than those.