chore/test: rename test_duplicate_votes_before_gossip
When looking at the output of the test:
test_duplicate_votes_before_gossip, one observes no duplicate vote. This
is because the behaviour of PARSEC is to only vote for a given event
once.

To make this obvious when casually browsing the test output, rename the
test to test_duplicate_vote_is_reduced_to_single.

MAID-3140