fix/parsec: prevent spurious InvalidGossipCreator accusations
This commit fixes several isssues:

Issue #1
--------

Consensus handling (`handle_self_consensus` and `handle_peer_consensus`) happened after the new meta election was created. Because of this, when the payload being decided was `Add(new_peer)`, the `new_peer` would not get added to the `undecided_peers` list of the new meta-election. This would cause problems in situation like this:

   1. There is Alice, Bob, Carol and Dave
   2. Alice decides `Add(Eric)`
   3. some time later, Alice decides `Add(Fred)`. The meta-election that decided it would not have Eric in `undecided_peers`.
   4. Alice detects consensus on `Add(Fred)` by Bob, Carol and Dave. That is all the peers in `undecided_peers`, so the meta-election is removed.
   5. Thus Alice never detects that Eric also decided `Add(Fred)`
   6. So when Alice receives the first event from Eric that has some event from Fred among its ancestors, she incorrectly accuses him of `InvalidGossipCreator`

This issue is fixed by moving the consensus handling before meta-election creation and by some supporting minor changes in `MetaElection` itself.

Issue #2
--------

Consider this example:

    1. Alice decides `Remove(Bob)`
    2. Alice then processes event from Eric and detects that Eric also decided `Remove(Bob)`. This event also has some events from Bob among its ancestors.
    3. Alice updates Eric's membership list to remove Bob
    4. Alice incorreclty accuses Eric of `InvalidGossipCreator`

The fix is that when the event under detection is also the event that decided `Remove`, we exclude that `Remove` from the membership list snapshot that is used for the detection. This fix required rather extensive change to membership list initialisation and handling.