fix/parsec: fix bug when parsing observations cache
The observations cache, when populated via parsing a dot file, wasn't
correctly flagging entries for which the 'created_by_us' field should
have been true.

This change partially fixes that, covering the case where the dot file
is complete, by moving the loop examining the events after the one
covering the metadata, and while iterating the events, setting the
'created_by_us' value correctly.

This still leaves the case where, when parsing a partial dot file, we
don't have all the events available to refer to.  So having defaulted to
setting the 'created_by_us' to false when examining the metadata, we
don't have the ability to update the value later.