chore/parsec: fix performance regression
When we started evaluating wether events had interesting ancestors to
resolve an issue with votes becoming "unseen" due to forks, we
introduced a performance regression as checking wether an event has an
interesting ancestor is expensive.

Regain the performance in the happy case by only performing this check
when this event sees forks.